KUDU-2113 Segfault because of consensus conflict and missing tablet server
Sometimes the master doesn't report all tablet servers. This can
happen, for example, when the master is starting. When this
happened and there was also a consensus conflict, it caused a
segfault. This fixes the segfault and contains a regression
test.

Change-Id: If80dbe687e70bd21de9dba9e81d71d66a5bd75e0
Reviewed-on: http://gerrit.cloudera.org:8080/7864
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>
(cherry picked from commit b54eab1f31dd2180e103f1cddbaa2194396396c5)
Reviewed-on: http://gerrit.cloudera.org:8080/10073
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>