consensus: Save previous last-logged OpId across tablet copies
When a replica falls behind the leader's log, and then fails a tablet
copy, it should still be able to vote in leader elections while
tombstoned. With this patch, that becomes possible. This helps with
availability.

Without this patch, asking a post-tablet copy failure replica to vote
would result in the following RPC response:

IllegalState: must be running to vote when last-logged opid is not known

Change-Id: Ie886764adbdc6d84406df44d6cdb693d038bcd57
Reviewed-on: http://gerrit.cloudera.org:8080/7888
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>
Reviewed-on: http://gerrit.cloudera.org:8080/7912
Reviewed-by: Dan Burkert <danburkert@apache.org>