tablet copy: Allow voting from failed initial tablet copies
This patch changes the behavior of tablets being copied for the first
time, i.e. to a new tablet server that has never hosted that replica. In
that case, if the tablet copy fails, with this patch that replica will
still be allowed to vote while tombstoned. This improves availability.
Without this patch, if a tablet copy to a new replica fails, that
replica will not be able to vote while tombstoned because its
last-logged OpId field in its superblock will be unset.
This is achieved within tablet copy by setting the initial
tombstoned_last_logged_opid to (1,0).

This patch is a bit of a hack, since it takes advantage of the fact that
it is impossible to have a "real" OpId of 1.0 in the current
implementation (1.1 is the operational minimum) so having a last-logged
OpId of 1.0 represents having never logged any actual ops.

Change-Id: Ibaad2f56797db731a0f701bbd37e3e450bc17b51
Reviewed-on: http://gerrit.cloudera.org:8080/7904
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins