master: always use smart pointers when accessing TableInfo and TabletInfo
Mixing scoped_refptr and raw access is error-prone and complicated code by
forcing the use of both in certain places. It's fair to be concerned about
the overhead; as far as I can tell it's minimal, because it's either in
functions that are already "slow" (like {Create,Alter,Delete}Table and the
equivalent in CatalogManagerBgTasks), or the call sites pass the object by
cref and thus avoid an extra ref/deref in the first place.

The one exception is in TableInfo::tablet_map, which would create a cycle if
it held strong refs to its TabletInfos. I tried to prevent this exception
from leaking outside of TableInfo.

I also reduced the number of ways to add/remove TabletInfo objects to a
TableInfo down to one.

Finally, I couldn't help myself and did some more modernization:
- Removed a couple std:: prefixes.
- Added "auto" to some loops.
- Replaced some vector::push_back calls with emplace_back.
- Reformatted double '>' (C++11 allows them to be adjacent).

Change-Id: I3c74f2aa048ee85cc3a5f863ce8f38147e78c5ae
Reviewed-on: http://gerrit.cloudera.org:8080/7909
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>