[periodic-test] fix flaky PeriodicTimerTest.TestReset
Fixed flake in PeriodicTimerTest.TestReset. The test was a little bit
flaky because sometimes the main test thread might have slept for much
longer than requested, sometimes longer than the period of the task.
If that happened, the test failed with messages like the following:

  src/kudu/rpc/periodic-test.cc:116: Failure
        Expected: 0
  To be equal to: counter_
        Which is: 1

This fix just increases the period of the task across the board.

Prior to this fix, if running a DEBUG or ASAN build with
--stress-cpu-threads=64, there were about 2 failures every 1K run:
  http://dist-test.cloudera.org//job?job_id=aserbin.1504248200.5336

After the fix there are no failures anymore:
  http://dist-test.cloudera.org//job?job_id=aserbin.1504299298.23673

Change-Id: I6848bee0e830075aba3ae5d8ed49f99038043063
Reviewed-on: http://gerrit.cloudera.org:8080/7932
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins