cow_object: lazily copy dirty state
This addresses a long-standing TODO that ought to improve performance of
tablet reports in the steady-state: within the COWed objects, defer the
creation of the "dirty" copy until it's actually needed (i.e. the first call
to mutable_dirty()).

The deferral introduces new branches in dirty() and mutable_dirty(), but
given the paucity of those calls for any given lock instance, I think this
is a reasonable trade-off.

Change-Id: Ic00c573717e7cb8df1942ba0760726fd0961e1d6
Reviewed-on: http://gerrit.cloudera.org:8080/7951
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>