test_workload: make table creation more robust
The table creation here already works around the lack of Exactly Once
semantics by considering certain kinds of failures to mean success. However,
these failures imply that the client didn't finish waiting for the table to
be created, so we need to do that ourselves.

Without the patch, ClientStressTest_MultiMaster.TestLeaderResolutionTimeout
failed 1/1000 times in DEBUG mode. With it, there were no failures.

Change-Id: Ic1e76e502359b499466cfa90d21ac22f35928261
Reviewed-on: http://gerrit.cloudera.org:8080/7982
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Hao Hao <hao.hao@cloudera.com>