Avoid unnecessary vector<Slice> allocations for ReadV/WriteV-like APIs
This reduces extra short-lived allocations for the various
scatter-gather (aka "vectored") read/write APIs.

These APIs used to all take vector<Slice>, which forces the caller to
allocate on the heap. Almost all of the call sites only need a small
constant number of slices, so the heap allocation is unnecessary
overhead.

This patch imports ArrayView<T> from Chromium/WebRTC. This is a view
into an existing array or vector (similar to how a StringPiece is a view
into a string or character array). The vectored read/write APIs are
converted to use ArrayView<Slice> instead of vector<Slice>.

Change-Id: I4eab29dad2e16cc5fce724d3bdd173f3a60cb266
Reviewed-on: http://gerrit.cloudera.org:8080/8077
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins