catalog_manager: make ScopedTabletInfoCommitter generic
I want to use it for TableInfo mutations, and to manage multiple read locks
too. So I've moved it to util/cow_object, created a read-only variant, and
templatized them for any CowObject.

Importantly, the default out-of-scope behavior is now Unlock(). Previously it
was Commit(), but I think Unlock() is less surprising.

Change-Id: I890db0fd18f773ab2253c49817def8162e04ad25
Reviewed-on: http://gerrit.cloudera.org:8080/8089
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>