Link to troubleshooting docs for two common problems
From experience, users don't know what to do when they see timeouts or
logs with "Soft memory limit exceeded" messages. Let's add a link to
some new troubleshooting docs, to give users an easy way to find solutions.

Likewise, when a disk failure or unexpected change to Kudu's data dirs
happens, users often don't know what to do when seeing the
"FsManager root not empty" message, so let's add a link to some new
troubleshooting docs for that, too.

The links go to the proper version of the docs, so a 1.6 server will
provide a link to the 1.6 docs. I added a couple of helper functions to
make these links, and put them in a new file since they didn't seem to
fit into any existing place.

Change-Id: Ida7e1495e4ba68f9e9b4d424650d84c0019c9b0f
Reviewed-on: http://gerrit.cloudera.org:8080/8093
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>