parse_metrics_log: merge metrics across entities of same type
This changes the tool so that all of the tablet metrics get summed
together before reporting. Previously we just reported the info from an
arbitrary tablet (probably the last one to be listed in the log).

This also adds a short circuit such that we skip parsing and merging of
metrics which have not been slated for output. This sped up processing
a metrics file ~10x when reporting only one or two metrics from it.

Lastly, this adds reporting of the max value from histograms.

Change-Id: I442cddd4c5352dcfae30fa23888082d916069d5f
Reviewed-on: http://gerrit.cloudera.org:8080/8062
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins