KUDU-2191 (1/n): Hive Metastore Kudu plugin
Adds a new plugin for the Hive Metastore whose purposes is to ensure
that Kudu table entries in the HMS are consistent. See the class
documentation for more specific information about what consistency
parameters are checked and enforced.

The plugin will live in the Kudu repository for the forseeble future
since it makes it easier to test against. We should consider upstreaming
it into Apache Hive eventually.

Change-Id: Ief0d64f71f90e04588a05098a44658fd461b9ec8
Reviewed-on: http://gerrit.cloudera.org:8080/8121
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins